refactor: file structure and post to delete for delete posts
