+++ server/src/main/java/org/example/Dorms/Dorm.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.Dorms;
 
 import java.util.List;
 import java.util.Set;
