+++ server/src/main/java/org/example/Dorms/DormDataSourceFactory.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.Dorms;
 
 import java.io.IOException;
