+++ server/src/main/java/org/example/Dorms/DormRoomTypesParser.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.Dorms;
 
 import java.io.*;
 import java.nio.file.*;
