+++ server/src/main/java/org/example/Dorms/FirebaseDormDatasource.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.Dorms;
 
 import com.google.auth.oauth2.GoogleCredentials;
 import com.google.cloud.firestore.CollectionReference;
