+++ server/src/main/java/org/example/Handlers/MatchHandler.java
@@ -3,8 +3,10 @@ package org.example.Handlers;
 
 import com.google.gson.JsonObject;
 import java.util.List;
-import org.example.DormDataSource;
-import org.example.DormSearchResult;
+
+import org.example.Dorms.DormDataSource;
+import org.example.Dorms.DormSearchResult;
+
 import spark.Request;
 import spark.Response;
 import spark.Route;
